chore(tests): unify structure — move component integration and stream tests to src/__tests__, keep unit tests colocated
